refactoring of file hierarchy
