+++ Sources/Swift/Integrations/AppHangTracking/SentryANRTrackerV2Delegate.swift
